==> 04 - Enriquecimiento de Geolocalización con Coordenadas.ipynb <==
# Importar Bibliotecas
import pyspark.sql.functions as F
import os
import pandas as pd
import time
from geopy.geocoders import Nominatim
from pyspark.sql import SparkSession

"""Crea una sesión de Spark con el nombre de la aplicación "Enriquecimiento de Geolocalización con Coordenadas", 
habilitando el soporte para Hive, y la inicializa o reutiliza si ya existe"""
spark = SparkSession \
    .builder \
    .appName("Enriquecimiento de Geolocalización con Coordenadas") \
    .enableHiveSupport() \
    .getOrCreate()

# Leer el archivo parquet de geolocalización de Correios
geo = spark.read.parquet('/home/felipe/despliegue_analytica/files_parquet/geolocation_correios.parquet')
# Leer el archivo parquet que contiene coordenadas geográficas
coords_df = spark.read.parquet('/home/felipe/despliegue_analytica/files_parquet/geo_coords.parquet')

"""Se seleccionan las columnas 'cep_prefix', 'city', y 'uf' del DataFrame 'geo' y se eliminan los duplicados.
Luego, se realiza una unión interna (inner join) con el DataFrame 'coords_df' que contiene las coordenadas, 
utilizando la columna 'cep_prefix' como clave de unión para garantizar que solo se mantengan los registros 
que coinciden en ambos DataFrames."""
geo_final = geo.select('cep_prefix','city', 'uf')\
   .dropDuplicates()\
   .join(
        coords_df,
        on=["cep_prefix"],
        how="inner"
    )

geo_final.show()

"""Se convierten las columnas 'lat' y 'lon' del DataFrame 'geo_final' al tipo de dato 'double' 
para garantizar mayor precisión al trabajar con coordenadas geográficas, asegurando que los valores 
de latitud y longitud se almacenen como números de punto flotante."""
geo_final = geo_final\
.withColumn('lat', F.col('lat').cast('double'))\
.withColumn('lon', F.col('lon').cast('double'))

# Mostrar el esquema del DataFrame final
geo_final.printSchema()

# Escribir el DataFrame resultante en un archivo parquet con las coordenadas de geolocalización
geo_final.write.parquet(f'/home/felipe/despliegue_analytica/files_parquet/geolocation_correios_coords.parquet', mode='overwrite')
